gbot2.0/requirements.txt
line-bot-sdk==3.1.1
gradio==4.44.0
requests==2.31.0
flask==2.3.3
python-dotenv==1.0.1
gunicorn==22.0.0